requirements.txt
psycopg2-binary>=2.9.9
requests>=2.31.0
python-dotenv>=1.0.0
pandas>=2.0.0
openpyxl>=3.1.0
google-genai>=0.2.0
word2number>=1.1

